Saving code to begin refactor

+++ run.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-nasm -f elf64 out.asm;
-ld out.o;
-./a.out;
-echo $?;
+nasm -f elf64 out.asm
+if [ $? -eq 0 ]; then
+  ld out.o
+  if [ $? -eq 0 ]; then
+    ./a.out
+    echo
+    echo "Exit code:" $?
+  fi
 
-rm out.o;
-rm a.out;
+  rm out.o
+  rm a.out
+fi
